renames a package, and gives general mongodb connection
