+++ src/main/java/sai_adapa/projs/inv_management/controllers/UsersController.java
@@ -1,7 +1,7 @@
 package sai_adapa.projs.inv_management.controllers;
 
 import org.springframework.web.bind.annotation.*;
-import sai_adapa.projs.inv_management.entities.users.io.PreUsers;
+import sai_adapa.projs.inv_management.model.users.io.PreUsers;
 import sai_adapa.projs.inv_management.services.UsersService;
 
 @RestController
