+++ src/main/java/sai_adapa/projs/inv_management/controllers/VendorController.java
@@ -3,7 +3,7 @@ package sai_adapa.projs.inv_management.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import sai_adapa.projs.inv_management.entities.users.io.PreVendor;
+import sai_adapa.projs.inv_management.model.users.io.PreVendor;
 import sai_adapa.projs.inv_management.services.VendorService;
 
 @RestController
