+++ src/main/java/sai_adapa/projs/inv_management/model/orders/Order.java
@@ -1,0 +1,4 @@
+package sai_adapa.projs.inv_management.model.orders;
+
+public class Order {
+}
