+++ src/main/java/sai_adapa/projs/inv_management/model/users/Admin.java
@@ -1,5 +1,4 @@
-package sai_adapa.projs.inv_management.entities.users;
-
+package sai_adapa.projs.inv_management.model.users;
 
 import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
@@ -9,29 +8,36 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import java.util.UUID;
 
 
+@Builder
 @Entity
 @Getter
 @Setter
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Vendor {
+public class Admin {
+
 
     @Id
     @GeneratedValue(generator = "uuid")
     @GenericGenerator(name = "uuid", strategy = "uuid2")
     @Column(unique = true, nullable = false)
     @Type(type = "pg-uuid")
-    private Long vendor_id;
-
-    private String name;
-
-    @Column(unique = true)
+    private UUID admin_id;
+    @Column(unique = true, nullable = false)
+    @Setter
     private String email;
-    private String description;
+    @Setter
     private String passwdHash;
+    @Setter
+    @Column(unique = true, nullable = true)
     private String sessionToken;
 
+
+    public Admin(String email, String passwdHash) {
+        this.email = email;
+        this.passwdHash = passwdHash;
+    }
 }
