+++ src/main/java/sai_adapa/projs/inv_management/model/users/Vendor.java
@@ -1,4 +1,5 @@
-package sai_adapa.projs.inv_management.entities.users;
+package sai_adapa.projs.inv_management.model.users;
+
 
 import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
@@ -8,36 +9,29 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import java.util.UUID;
 
 
-@Builder
 @Entity
 @Getter
 @Setter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Admin {
-
+public class Vendor {
 
     @Id
     @GeneratedValue(generator = "uuid")
     @GenericGenerator(name = "uuid", strategy = "uuid2")
     @Column(unique = true, nullable = false)
     @Type(type = "pg-uuid")
-    private UUID admin_id;
-    @Column(unique = true, nullable = false)
-    @Setter
+    private Long vendor_id;
+
+    private String name;
+
+    @Column(unique = true)
     private String email;
-    @Setter
+    private String description;
     private String passwdHash;
-    @Setter
-    @Column(unique = true, nullable = true)
     private String sessionToken;
 
-
-    public Admin(String email, String passwdHash) {
-        this.email = email;
-        this.passwdHash = passwdHash;
-    }
 }
