+++ src/main/java/sai_adapa/projs/inv_management/model/users/io/PreUsers.java
@@ -1,4 +1,4 @@
-package sai_adapa.projs.inv_management.entities.users.io;
+package sai_adapa.projs.inv_management.model.users.io;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
