+++ src/main/java/sai_adapa/projs/inv_management/repositories/AdminRepository.java
@@ -2,7 +2,7 @@ package sai_adapa.projs.inv_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sai_adapa.projs.inv_management.entities.users.Admin;
+import sai_adapa.projs.inv_management.model.users.Admin;
 
 @Repository
 public interface AdminRepository extends JpaRepository<Admin,String> {
