+++ src/main/java/sai_adapa/projs/inv_management/repositories/ItemRepository.java
@@ -2,8 +2,8 @@ package sai_adapa.projs.inv_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sai_adapa.projs.inv_management.entities.items.Item;
+import sai_adapa.projs.inv_management.model.items.Item;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item,String> {
+public interface ItemRepository extends JpaRepository<Item,Long> {
 }
