+++ src/main/java/sai_adapa/projs/inv_management/repositories/UsersRepository.java
@@ -2,7 +2,7 @@ package sai_adapa.projs.inv_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sai_adapa.projs.inv_management.entities.users.Users;
+import sai_adapa.projs.inv_management.model.users.Users;
 
 @Repository
 public interface UsersRepository extends JpaRepository<Users,String> {
