+++ src/main/java/sai_adapa/projs/inv_management/repositories/VendorRepository.java
@@ -2,7 +2,7 @@ package sai_adapa.projs.inv_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sai_adapa.projs.inv_management.entities.users.Vendor;
+import sai_adapa.projs.inv_management.model.users.Vendor;
 
 @Repository
 public interface VendorRepository extends JpaRepository<Vendor,String > {
