+++ src/main/java/sai_adapa/projs/inv_management/services/UsersService.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import sai_adapa.projs.inv_management.tools.AuthTools;
 import sai_adapa.projs.inv_management.repositories.UsersRepository;
-import sai_adapa.projs.inv_management.entities.users.Users;
+import sai_adapa.projs.inv_management.model.users.Users;
 
 @Service
 public class UsersService {
