+++ src/main/java/sai_adapa/projs/inv_management/services/VendorService.java
@@ -2,7 +2,7 @@ package sai_adapa.projs.inv_management.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import sai_adapa.projs.inv_management.entities.users.Vendor;
+import sai_adapa.projs.inv_management.model.users.Vendor;
 import sai_adapa.projs.inv_management.repositories.VendorRepository;
 import sai_adapa.projs.inv_management.tools.AuthTools;
