+++ src/main/resources/application.properties
@@ -1,3 +1,4 @@
 logging.level.org.zalando.logbook= TRACE
 logbook.format.style=json
 logging.file.name= X:/spring_logs/inv.logs
+spring.data.mongodb.uri=mongodb://localhost:27017/inv_man
